🛠️ refactor: 파일 구조 변경 🔗 Resolves: #44
