+++ backend/JiShop/src/main/java/com/jishop/member/controller/NaverControllerImpl.java
@@ -3,8 +3,8 @@ package com.jishop.member.controller;
 import com.jishop.member.domain.LoginType;
 import com.jishop.member.domain.User;
 import com.jishop.member.dto.SocialUserInfo;
-import com.jishop.service.OauthService;
-import com.jishop.service.UserService;
+import com.jishop.member.service.OauthService;
+import com.jishop.member.service.UserService;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.ResponseEntity;
