+++ backend/JiShop/src/main/java/com/jishop/member/controller/UserControllerImpl.java
@@ -4,7 +4,7 @@ import com.jishop.member.dto.FinalStepRequest;
 import com.jishop.member.dto.SignUpFormRequest;
 import com.jishop.member.dto.Step1Request;
 import com.jishop.member.dto.Step2Request;
-import com.jishop.service.UserService;
+import com.jishop.member.service.UserService;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
