+++ backend/JiShop/src/main/java/com/jishop/member/service/EmailSenderServiceImpl.java
@@ -1,8 +1,7 @@
-package com.jishop.service.impl;
+package com.jishop.member.service;
 
 import com.amazonaws.services.simpleemail.AmazonSimpleEmailService;
 import com.amazonaws.services.simpleemail.model.*;
-import com.jishop.service.EmailSender;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
