+++ backend/JiShop/src/main/java/com/jishop/member/service/GoogleServiceImpl.java
@@ -1,8 +1,7 @@
-package com.jishop.service.impl;
+package com.jishop.member.service;
 
 import com.jishop.member.dto.SocialUserInfo;
 import com.jishop.member.dto.TokenResponse;
-import com.jishop.service.AbstractOAuthService;
 import jakarta.servlet.http.HttpSession;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
